changed forage source bar value type to byte

+++ Entity/IForageSourceEntity.cs
@@ -41,6 +41,6 @@
         /// <summary>
         /// Get the current value for a specific bar
         /// </summary>
-        float GetCurrentBarValue(TfsBarIndex index);
+        byte GetCurrentBarValue(TfsBarIndex index);
     }
 }
